requirements.txt
py-clob-client>=0.16.0
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6
tabulate>=0.9.0
web3>=6.0.0
